fix multiple nestjs instances
