+++ back/src/auth/notif/notif.gateway.ts
@@ -7,9 +7,10 @@ import {
   WebSocketGateway,
   WebSocketServer,
 } from '@nestjs/websockets';
-import { Socket, Server } from 'socket.io';
+import { Socket, Server, Namespace } from 'socket.io';
 import { SocketService } from '../../sockets/socket.service';
 import { NotifService } from './notif.service';
+import { Cron } from '@nestjs/schedule';
 
 @WebSocketGateway({
   cors: {
@@ -25,22 +26,36 @@ export class NotifGateway
     private notifService: NotifService,
   ) {}
   @WebSocketServer()
-  server: Server;
+  server: Namespace;
 
   afterInit(server: Server) {
+    console.log('notifGateway running')
     this.notifService.server = server;
+
+    // this is debug, not necessary for production
+    server.use((client: Socket, next) => {
+      client.use((event, next) => {
+        console.log(
+          '\x1b[36m%s\x1b[0m',
+          'Middleware: New socket event',
+          event[0],
+        );
+        next();
+      });
+      next();
+    });
   }
 
   // handle connection
   handleConnection(client: Socket) {
-    // console.log(`client with id: ${client.id} is connected !`);
+    console.log(`client with id: ${client.id} is connected (notif) !`);
     this.socketService.handleConnection(client, '');
   }
 
   // handle disconnect
   handleDisconnect(client: Socket) {
     // this.notifService.handleDisconnect(client, userLogin, '');
-    console.log(`client with id: ${client.id} has left the connection !`);
+    console.log(`client with id: ${client.id} has left the connection (notif) !`);
   }
 
   @SubscribeMessage('client.notif.chatNotif')
@@ -64,6 +79,11 @@ export class NotifGateway
   // handleFriendsRequestNotif() {
   //   this.server.emit('friendsNotif');
   // }
+
+  @Cron('*/5 * * * * *')
+  private debug() {
+    console.log('[Debug NotifGateway]', { rooms: this.server.adapter.rooms });
+  }
 }
 
 // https://docs.nestjs.com/websockets/gateways
